+++ src/components/Atoms/Grades/PeriodSelector.jsx
@@ -1,9 +1,7 @@
 import React from 'react'
 
-import DropdownButton from 'cozy-ui/transpiled/react/DropdownButton'
-import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
-import Menu from 'cozy-ui/transpiled/react/Menu'
-import MenuItem from 'cozy-ui/transpiled/react/MenuItem'
+import { PeriodSelectorButton } from './PeriodSelectorButton'
+import { YearSelectorButton } from './YearSelectorButton'
 
 export const PeriodSelector = ({
   periodDropdownRef,
@@ -51,122 +49,5 @@ export const PeriodSelector = ({
         t={t}
       />
     </div>
-  )
-}
-
-export const PeriodSelectorButton = ({
-  periodDropdownRef,
-  selectedPeriod,
-  setSelectedPeriod,
-  setPeriodMenuOpen,
-  periodMenuOpen,
-  periods,
-  t,
-  textVariant
-}) => {
-  return (
-    <>
-      <DropdownButton
-        ref={periodDropdownRef}
-        aria-controls="simple-menu"
-        aria-haspopup="true"
-        onClick={() => setPeriodMenuOpen(!periodMenuOpen)}
-        textVariant={textVariant}
-      >
-        {selectedPeriod || t('Grades.selectPeriod')}
-      </DropdownButton>
-
-      <Menu
-        open={periodMenuOpen}
-        anchorEl={periodDropdownRef.current}
-        getContentAnchorEl={null}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left'
-        }}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'left'
-        }}
-        keepMounted
-        onClose={() => setPeriodMenuOpen(false)}
-      >
-        {periods.map((period, i) => (
-          <MenuItem
-            key={i}
-            onClick={() => {
-              setSelectedPeriod(period)
-              setPeriodMenuOpen(false)
-            }}
-            selected={period === selectedPeriod}
-          >
-            <ListItemText primary={period} />
-          </MenuItem>
-        ))}
-
-        {periods.length === 0 && (
-          <MenuItem disabled>
-            <ListItemText primary={t('Grades.emptyList.periods')} />
-          </MenuItem>
-        )}
-      </Menu>
-    </>
-  )
-}
-
-export const YearSelectorButton = ({
-  yearDropdownRef,
-  selectedYear,
-  selectedPeriod,
-  setSelectedYear,
-  setYearMenuOpen,
-  yearMenuOpen,
-  allPeriods,
-  t
-}) => {
-  return (
-    <>
-      <DropdownButton
-        ref={yearDropdownRef}
-        aria-controls="simple-menu"
-        aria-haspopup="true"
-        onClick={() => setYearMenuOpen(!yearMenuOpen)}
-      >
-        {selectedYear || t('Grades.selectYear')}
-      </DropdownButton>
-
-      <Menu
-        open={yearMenuOpen}
-        anchorEl={yearDropdownRef.current}
-        getContentAnchorEl={null}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left'
-        }}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'left'
-        }}
-        keepMounted
-        onClose={() => setYearMenuOpen(false)}
-      >
-        {[
-          ...new Set(
-            allPeriods.filter(p => p.title === selectedPeriod).map(p => p.year)
-          )
-        ].map((year, i) => (
-          <MenuItem
-            key={i}
-            onClick={() => {
-              setSelectedYear(year)
-              setYearMenuOpen(false)
-            }}
-            selected={year === selectedYear}
-          >
-            <ListItemText primary={year} />
-          </MenuItem>
-        ))}
-      </Menu>
-    </>
   )
 }
